Removed unnecessary code and refactored a bit.
- Changed the structure of the project so it runs smoothly and has less unused assets inside.
