--- backend/remindit-backend/src/main/java/com/remindit/reminditbackend/ReminditApplication.java
@@ -1,13 +1,0 @@
-package com.remindit.reminditbackend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReminditApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ReminditApplication.class, args);
-	}
-
-}
