--- backend/remindit-backend/src/main/resources/remindit-87ab4-firebase-adminsdk-bctjs-a110238b99.json
@@ -1,13 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "remindit-87ab4",
-  "private_key_id": "a110238b99cb9fa7bcfee834d1b029db4f003f6a",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCioFDwIuRsLAlt\nC/gyO+uo4gMUUILnYedpjMlhXeRwvxBCK5011D6wZfrZCUDWfHkmnEroM1tovmUW\n9IWBshURNNHcJ+ljU0GvktB1YicI/BauoXQhE8QRLa3BdrBVUQg0eraj838+W6hd\nuk8U/Xf4ULsX4PtFKuN1IjfR531fkLbxIGUAvZ0gDuUw6+8RnBljL7DKPd1BQdZT\nQuKx+Fm3uAD0hMNKwrodd0vIFTO6jjUyRzbtkgHJ13hxizI4wJR4KeJ838TobhpY\npo3FGwyXgbxpYJHFPe26VWhnl+RDVcvyWy8qcOnEdBCnelCwCy1QWgb0GoRdD665\nE3kgrrAvAgMBAAECggEAE1IMi5r5nEhXDld/sygfRT8ksoMmKisdfRK20hR/kU2B\ntvrLgJuonvhepu5Mf67mDco/s/p1/6mvBN0kB1aK3jJyRD+RoLVBdoRws6Fnlc+t\nOi4AR/W7M3dDz1o0dM5dtS5nCG9jTJtDE2+Q+aNrX1rn/1xBPeRshXv9ObjQ0kuE\ncTul8MacUe/ILzW2TncYtRyQA/22If2aMYHZlNJQ+YTYuzdi3goepJblHwnXsufj\nO6XHsOHqLV4MqbV6GfCI5fBALE9vee5zSQE/IgCIt9asV7Op4a7klbyAj4in+vxc\nWDjcBpDcw8Tc0iIkwLUoOtYiaiVD/ZReG/7oUfQMwQKBgQDcKEZOA3FtXUTJFg0O\nY1iGHOp2rGjG5W4RD431dDPginO+zqKPbGvd27n+0Kf4qTe8qbINxo9u0wRBwldP\nAwmUb5ofEsTV+e4oCjadHto11Mi8GSIhTndvai/NsIoACmfiAgYurRtCrnsL3T7R\nxLpcZI0oCLhwyEZVxJhB+8rRkQKBgQC9GkI2q96pZavbTLgpVrFJ4WuHCVMT3qF1\ndlb8rC4npBlgEPfmfS5DvPyzrzEJdfZ1QsdsI5qCJ0gyrlfdLd5cvPd2MeTi6yYD\njZ0+VegbNBx9HJ9ylvjKT1JxOEvrErZNvevHftFeEpDFbmarm3y2WsoiLoibWIoY\n9aGsj7yFvwKBgCZhOwhkn0SCivhHSNVmSktvBQzVN5rc6lg9DvFoZZbSbCUqXrqE\nGzXdFUqv02iND3NrXef64oxeaxKPr6rNi8D9x699FSV8YaFHTogLldzmkGwCMan+\nxOF+cID6MsSZC80hrzQ551OzmS3TTx1An4B14wnEo2q9700KbcSrAYuRAoGAeoem\nOciZyUeg2MNYXSjYTZfr1Z7cHvbVxce9S5HBglwZOR8lkkqwjKHLWB5/drLh1E9F\n6rMTLZyIcpZ3RlCvUYX8DrD7r3FQu2Kqbba6Y84bW5INWvGqGWTWdZhV0ObaC3ux\naB2rSfpbpiIIEs/4OiyP5yAh0g7HxK58aNTmAf8CgYEAmFR/0gRSkmpRVDeDnP3+\nbxthoVvzQSeFD+mEXbrdJlfJqiEuo9eKk5RpSenphFVDu3QCdW//NDY2FaSa3zIB\nr9tkhssyr/uMnsamjiAJF6N4vacRShLZdPsPSxrey03MR80JubqLyMPTIDD8QEpB\nc1f64dP3ESE8Xsjt9wdlovE=\n-----END PRIVATE KEY-----\n",
-  "client_email": "firebase-adminsdk-bctjs@remindit-87ab4.iam.gserviceaccount.com",
-  "client_id": "112679809585408502828",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-bctjs%40remindit-87ab4.iam.gserviceaccount.com",
-  "universe_domain": "googleapis.com"
-}
