--- backend/remindit-backend/src/test/java/com/remindit/remindit/ReminditApplicationTests.java
@@ -1,13 +1,0 @@
-package com.remindit.remindit;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReminditApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
